requirements.txt
Flask==0.10
Flask-WTF==0.9.1
py-bcrypt==0.4
PIL==1.1.7
